Intermidiate refactoring step
