+++ main.py
@@ -2,4 +2,4 @@ from src.controller import Controller
 
 if __name__ == "__main__":
     controller = Controller(window_name="TelegramDesktop")
-    controller.run(10)
+    controller.run(seconds_to_play=10)
