+++ tests/test_normalize_point.py
@@ -1,13 +1,13 @@
-import sys
 import os
+import sys
 
 script_dir = os.path.dirname(os.path.abspath(__file__))
 parent_dir = os.path.dirname(script_dir)
 sys.path.append(parent_dir)
 
-from src.controller import Controller
-
 import numpy as np
+
+from src.controller import Controller
 
 ct = Controller()
